+++ Codigo/php/horarios_aulas_cursos-index.php
@@ -24,22 +24,26 @@
     <div class="container-fluid">
         <div class="row">
             <div class="col-md-12">
-            <h1>Detallo de Horarios-Curso</h1>
-        <div class="d-flex justify-content-end align-items-center mb-5">
-            <a href="horarios_aulas_cursos-create.php" class="btn btn-success mr-3"><i class='bx bx-sm bx-plus'></i> Nuevo registro</a>
-            <a href="horarios_aulas_cursos-index.php" class="btn btn-info mr-3">Actualizar</a>
-            <a href="index.php" class="btn btn-secondary"><i class='bx bx-sm bx-arrow-back'></i> Atrás</a>
-        </div>
+                <h1>Detallo de Horarios-Curso</h1>
+                <div class="d-flex justify-content-end align-items-center mb-5">
+                    <a <?php if (!in_array("crear", $privilegios)) {
+                        echo 'style="display: none;"';
+                    } ?>
+                        href="horarios_aulas_cursos-create.php" class="btn btn-success mr-3"><i
+                            class='bx bx-sm bx-plus'></i> Nuevo registro</a>
+                    <a href="horarios_aulas_cursos-index.php" class="btn btn-info mr-3">Actualizar</a>
+                    <a href="index.php" class="btn btn-secondary"><i class='bx bx-sm bx-arrow-back'></i> Atrás</a>
+                </div>
 
                 <div class="form-row">
-            <form action="horarios_aulas_cursos-index.php" method="get">
-                <div class="d-flex">
-                    <input type="text" class="form-control mr-2" placeholder="Buscar en la tabla"
-                        aria-label="Buscar en la tabla" name="search" autofocus>
-                        <button type="submit" class="btn btn-primary"><i class='bx bx-search-alt-2'></i></button>
+                    <form action="horarios_aulas_cursos-index.php" method="get">
+                        <div class="d-flex">
+                            <input type="text" class="form-control mr-2" placeholder="Buscar en la tabla"
+                                aria-label="Buscar en la tabla" name="search" autofocus>
+                            <button type="submit" class="btn btn-primary"><i class='bx bx-search-alt-2'></i></button>
+                        </div>
+                    </form>
                 </div>
-            </form>
-        </div>
                 <br>
 
                 <?php
@@ -149,13 +153,17 @@
                         $number_of_results = mysqli_num_rows($result_count);
                         echo " " . $number_of_results . " Resultado - Página " . $pageno . " de " . $total_pages;
 
+                        echo "<div class='card shadow mb-4 p-1'>";
+                        echo "<div class='card-body'>";
+                        echo "<div class='table-responsive'>";
+                        echo "<table class='table table-bordered table-striped'>";
                         echo "<table class='table table-bordered table-striped'>";
                         echo "<thead>";
                         echo "<tr>";
                         echo "<th><a href=?search=$search&sort=&order=id_horarios_aulas_cursos&sort=$sort>ID</th>";
                         echo "<th><a href=?search=$search&sort=&order=id_curso&sort=$sort>Curso</th>";
                         echo "<th><a href=?search=$search&sort=&order=id_horario__aula&sort=$sort>Horario</th>";
-                        
+
 
                         echo "<th>Acción</th>";
                         echo "</tr>";
@@ -167,14 +175,27 @@
                             echo "<td>" . htmlspecialchars($row['id_curso']) . "</td>";
                             echo "<td>" . htmlspecialchars($row['id_horario__aula']) . "</td>";
                             echo "<td>";
-                            echo "<a href='horarios_aulas_cursos-read.php?id_horarios_aulas_cursos=" . $row['id_horarios_aulas_cursos'] . "' title='Ver Registro' data-toggle='tooltip'><i class='far fa-eye'></i></a>";
-                            echo "<a href='horarios_aulas_cursos-update.php?id_horarios_aulas_cursos=" . $row['id_horarios_aulas_cursos'] . "' title='Actualizar Registro' data-toggle='tooltip'><i class='far fa-edit'></i></a>";
-                            echo "<a href='horarios_aulas_cursos-delete.php?id_horarios_aulas_cursos=" . $row['id_horarios_aulas_cursos'] . "' title='Eliminar Registro' data-toggle='tooltip'><i class='far fa-trash-alt'></i></a>";
+
+                            if (in_array("seleccionar", $privilegios)) {
+                                echo "<a href='horarios_aulas_cursos-read.php?id_horarios_aulas_cursos=" . $row['id_horarios_aulas_cursos'] . "' title='Ver Registro' data-toggle='tooltip'><i class='far fa-eye'></i></a>";
+                            }
+
+                            if (in_array("modificar", $privilegios)) {
+                                echo "<a href='horarios_aulas_cursos-update.php?id_horarios_aulas_cursos=" . $row['id_horarios_aulas_cursos'] . "' title='Actualizar Registro' data-toggle='tooltip'><i class='far fa-edit'></i></a>";
+                            }
+
+                            if (in_array("eliminar", $privilegios)) {
+                                echo "<a href='horarios_aulas_cursos-delete.php?id_horarios_aulas_cursos=" . $row['id_horarios_aulas_cursos'] . "' title='Eliminar Registro' data-toggle='tooltip'><i class='far fa-trash-alt'></i></a>";
+                            }
+
                             echo "</td>";
                             echo "</tr>";
                         }
                         echo "</tbody>";
                         echo "</table>";
+                        echo "</div>";
+                        echo "</div>";
+                        echo "</div>";
                         ?>
                         <ul class="pagination" align-right>
                             <?php
@@ -185,22 +206,20 @@
                             <li class="page-item <?php if ($pageno <= 1) {
                                 echo 'disabled';
                             } ?>">
-                                <a class="page-link"
-                                    href="<?php if ($pageno <= 1) {
-                                        echo '#';
-                                    } else {
-                                        echo $new_url . "&pageno=" . ($pageno - 1);
-                                    } ?>">Previa</a>
+                                <a class="page-link" href="<?php if ($pageno <= 1) {
+                                    echo '#';
+                                } else {
+                                    echo $new_url . "&pageno=" . ($pageno - 1);
+                                } ?>">Ant</a>
                             </li>
                             <li class="page-item <?php if ($pageno >= $total_pages) {
                                 echo 'disabled';
                             } ?>">
-                                <a class="page-link"
-                                    href="<?php if ($pageno >= $total_pages) {
-                                        echo '#';
-                                    } else {
-                                        echo $new_url . "&pageno=" . ($pageno + 1);
-                                    } ?>">Siguiente</a>
+                                <a class="page-link" href="<?php if ($pageno >= $total_pages) {
+                                    echo '#';
+                                } else {
+                                    echo $new_url . "&pageno=" . ($pageno + 1);
+                                } ?>">Sig</a>
                             </li>
                             <li class="page-item <?php if ($pageno >= $total_pages) {
                                 echo 'disabled';
